[Dist] move DistNode/EdgeDataLoader to distributed
