+++ python/dgl/dataloading/__init__.py
@@ -1,4 +1,5 @@
 """Package for dataloaders and samplers."""
+
 from .. import backend as F
 from . import negative_sampler
 from .base import *
@@ -11,4 +12,3 @@ from .shadow import *
 if F.get_preferred_backend() == "pytorch":
     from .spot_target import *
     from .dataloader import *
-    from .dist_dataloader import *
