+++ python/dgl/distributed/__init__.py
@@ -1,7 +1,12 @@
 """DGL distributed module"""
+
 from . import optim
 from .dist_context import exit_client, initialize
-from .dist_dataloader import DistDataLoader
+from .dist_dataloader import (
+    DistDataLoader,
+    DistEdgeDataLoader,
+    DistNodeDataLoader,
+)
 from .dist_graph import DistGraph, DistGraphServer, edge_split, node_split
 from .dist_tensor import DistTensor
 from .graph_partition_book import GraphPartitionBook, PartitionPolicy
