+++ tests/python/pytorch/test_dgl_import.py
@@ -1,0 +1,7 @@
+import sys
+
+
+def test_graphbolt_is_not_imported():
+    import dgl
+
+    assert "dgl.graphbolt" not in sys.modules
